pyproject.toml
[project]
name = "multi-agent-orchestration-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.10",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
]
